refactor: toggle icon moved to separate component
