+++ client/src/components/Messaging/UserStatusInfo.js
@@ -1,13 +1,8 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import styles from './styles/UserStatusInfo.module.css';
-import { ThemeContext } from '../../ThemeContext.js';
-import { FiSun, FiMoon } from 'react-icons/fi';
+import ThemeToggle from '../ThemeToggle/index.js';
 
 export const UserStatusInfo = ({ online }) => {
-  const [darkMode, setDarkMode] = useContext(ThemeContext);
-  const toggleTheme = () => {
-    setDarkMode(!darkMode);
-  };
   return (
     <div className={styles.userInfo}>
       {online ? (
@@ -17,13 +12,7 @@ export const UserStatusInfo = ({ online }) => {
       ) : (
         'Waiting for Alice to join..'
       )}
-      <span className={styles.toggleIcon}>
-        {darkMode === true ? (
-          <FiSun className={styles.FiSun} onClick={toggleTheme} />
-        ) : (
-          <FiMoon className={styles.FiMoon} onClick={toggleTheme} />
-        )}
-      </span>
+      <ThemeToggle />
     </div>
   );
 };
