+++ client/src/components/ThemeToggle/Style.module.css
@@ -1,0 +1,11 @@
+.toggleIcon {
+  font-size: 24px;
+}
+
+.FiMoon {
+  color: #231709;
+}
+
+.FiSun {
+  color: #4bb2f9;
+}
